Finalise class port and remove old driver file
