+++ c/src/permissions/accounts.h
@@ -2,7 +2,7 @@
 
 #include <stdint.h>
 #include "../deps/json.hpp"
-#include "../storage/driver.h"
+#include "../storage/table.h"
 #include "permissions.h"
 
 struct AccountPassword {
@@ -22,8 +22,8 @@ void create_database_account(char* username, char* password, DatabasePermissions
 void delete_database_account(DatabaseAccount* account);
 void update_database_account(DatabaseAccount* account, DatabaseAccount new_account);
 
-void set_table_account_permissions(active_table* table, DatabaseAccount* account, TablePermissions& permissions);
-void delete_table_account_permissions(active_table* table, DatabaseAccount* account);
+void set_table_account_permissions(ActiveTable* table, DatabaseAccount* account, TablePermissions& permissions);
+void delete_table_account_permissions(ActiveTable* table, DatabaseAccount* account);
 
-const TablePermissions* get_table_permissions_for_account(active_table* table, DatabaseAccount* account, bool include_table_admin = true);
+const TablePermissions* get_table_permissions_for_account(ActiveTable* table, DatabaseAccount* account, bool include_table_admin = true);
